requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
pandas==1.3.3
baostock==0.8.9
numpy==1.21.2
click==8.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
python-dotenv==0.19.0
tushare==1.2.89
mysql-connector-python==8.0.26
SQLAlchemy==1.4.23
requests==2.26.0
APScheduler==3.8.1
pytz==2021.1
python-dateutil==2.8.2